카테고리 수정기능 추가

+++ FE/src/components/admin.vue
@@ -1,23 +1,33 @@
 <template>
     <div id="admin">
 
-<!--대분류 카테고리 -->
+        <!--대분류 카테고리 -->
         <nav class="mainCategoryGroup">
             <br>
-            <h3>대분류 카테고리</h3>
+            <h3>Main Category</h3>
             <br>
 
             <!--대분류카테고리 추가-->
             <div class="createCategory">
                 <br>
-                <h4>카테고리 추가하기</h4>
+                <h4>Add Category</h4>
+
+                <div class="hrLine20">
+                    <hr>
+                </div>
+
                 <div class="inputMain">
-                    <label for="mainCategory"> 대분류 이름: </label>  &nbsp;
+                    <label for="mainCategory"> 카테고리 이름: </label>  &nbsp;
                     <input id="mainCategory" size="sm" type="text" placeholder="입력해주세요." v-model="addCategoryName"></input>
                 </div>
 
+                <div class="hrLine20">
+                    <hr>
+                </div>
+
                 <b-form-group>
-                    이미지 업로드 방식:
+                    <h6>이미지 업로드</h6>
+                    <br>
                     <b-form-radio-group
                             v-model="condition.imgUpload"
                             :options="uploadMethod">
@@ -45,28 +55,66 @@
 
 
                 <!--추가 버튼-->
-                <button class="plusbtn btn btn-primary" v-on:click="addCategory" >+</button>
+                <button class="plusbtn btn btn-primary" v-on:click="addCategory" >추가하기</button>
             </div>
 
 
 
-        <!--대분류카테고리 리스트-->
+            <!--대분류카테고리 리스트-->
 
             <div  v-for="item in mainCategoryList.slice().reverse()" :key="item._id">
-                <table class="categoryTable">
+                <!--카테고리 이름-->
+                <table class="categoryTable" v-on:click="categoryClicked(item.name,item._id)">
                     <tr>
                         <th class="categoryImg"  rowspan="2">
-                            <input v-on:click="categoryClicked(item.name,item._id)" type="image" :src="item.img_path"
-                                   height="140px" width="140px"/>
+                            <input  type="image" :src="item.img_path"
+                                    height="140px" width="140px"/>
                         </th>
                         <th class="mainCategoryTitle">
-                            <input class="mainCategoryTitle" name="mainCategoryTitleInput"  type="text" v-model="item.name">
+                            {{item.name}}
                         </th>
                     </tr>
                     <tr>
                         <td class="tg-0lax">
-                            <button class="btn btn-primary" v-on:click="categoryUpdate(item)">수정</button>
+                            <div class="hrLine20">
+                                <hr>
+                            </div>
+                            <button class="btn btn-primary" v-if="condition.reviseComplete=='revise'" v-on:click="revisebtnChange">수정</button>
+
                             <button class="btn btn-danger" v-on:click="categoryDelete(item._id)">삭제</button>
+                        </td>
+                    </tr>
+
+                    <tr v-if="condition.reviseComplete=='complete'&&item._id==selectedCategoryID">
+
+                        <!--수정 기능 -->
+                        <td colspan="2">
+                            <label for="reviseCategory"> 이름: </label>  &nbsp;
+                            <input  id="reviseCategory" class="mainCategoryTitle" name="mainCategoryTitleInput"  type="text" v-model="item.name">
+                            <br>
+                            <b-form-radio-group
+                                    v-model="condition.reviseimgUpload"
+                                    :options="uploadMethod">
+                            </b-form-radio-group>
+
+                            <!--파일 업로드 형식-->
+                            <div class="imgUpload" v-if="condition.reviseimgUpload=='methodFile'">
+                                <b-form-file class="file_input " v-model="file" accept=".jpg, .png" :state="Boolean(file)"
+                                             placeholder="Choose a file..."
+                                             @change="onFileChange($event.target.files)"></b-form-file>
+                                <b-row>
+                                    <div class="preview" >
+                                        <b-img v-if="file" :src="file.blob"/>
+                                    </div>
+                                </b-row>
+                            </div>
+
+                            <div class="imgUpload" v-if="condition.reviseimgUpload=='methodURL'">
+                                <input id="reviseimgURL" size="sm" type="text" placeholder="입력해주세요." v-model="reviseCategoryPath"/>
+                            </div>
+
+
+                            <button class="btn btn-primary" v-if="condition.reviseComplete=='complete'" v-on:click="categoryUpdate(item)">완료</button>
                         </td>
                     </tr>
                 </table>
@@ -76,48 +124,54 @@
 
 
 
-<!--대분류에 속해있는 소분류 카테고리 -->
+        <!--대분류에 속해있는 소분류 카테고리 -->
         <article class="subCategory">
             <br>
 
             <div class="subCategoryTitle" v-if="selectedCategoryID==-1">
-                <h2> 대분류 카테고리를 선택해주세요</h2>
+                <h2> 메인카테고리를 선택해주세요</h2>
             </div>
-
 
 
             <div class="subCategoryTitle" v-if="selectedCategoryID!=-1">
-                <h2>{{selectedCategoryName}}의 소분류 카테고리입니다.</h2>
+                <h2>{{selectedCategoryName}}의 세부 카테고리</h2>
             </div>
 
-    <!--소분류 카테고리 추가-->
-            <div class="createCategory" v-if="selectedCategoryID!= -1">
-                <div class="inputMain">
-                    <label for="subCategory"> 소분류 이름: </label>  &nbsp;
-                    <input id="subCategory" size="sm" type="text" placeholder="입력해주세요." v-model="addSubCategoryName"></input>
+            <!--소분류 카테고리 추가-->
+            <div class="createsubCategoryArea">
+                <div class="createsubCategory" v-if="selectedCategoryID!= -1">
+                    <div class="inputMain">
+                        <label for="subCategory"> 소분류 이름: </label>  &nbsp;
+                        <input id="subCategory" size="sm" type="text" placeholder="입력해주세요." v-model="addSubCategoryName"></input>
+                    </div>
+
+
+                    <!--추가 버튼 -->
+                    <button class="subplusbtn btn btn-primary" v-on:click="addSubCategory" >추가하기</button>
                 </div>
-
-
-                <!--추가 버튼 -->
-                <button class="subplusbtn btn btn-primary" v-on:click="addSubCategory" >+</button>
             </div>
-
             <b-container class="subCategoryGroup" >
 
 
-    <!--소분류 카테고리 리스트-->
+                <!--소분류 카테고리 리스트-->
                 <div v-for="item in mainCategoryList" :key="item.id">
                     <b-row v-if="selectedCategoryID==item._id">
                         <div  v-for="sub in item.sub_category.slice().reverse()">
                             <b-col class="subCategorySingle">
                                 <input class="subCategoryListTitle" name="subCategoryTitleInput"  type="text" v-model="sub.name"> <br>
-                                <button class="btn btn-primary" v-on:click="subcategoryUpdate(sub)">수정</button>
+                                <div class="hrLine20">
+                                    <hr>
+                                </div>
+                                <button class="btn btn-primary"  v-on:click="subcategoryUpdate(sub)">수정</button>
                                 <button class="btn btn-danger" v-on:click="subCategoryDelete(sub._id)">삭제</button>
                             </b-col>
                         </div >
                     </b-row>
                 </div>
             </b-container>
+
+
+
 
         </article>
 
@@ -147,10 +201,14 @@
                 formData : new FormData(),
                 addCategoryName:"",
                 addCategoryPath:"",
+                reviseCategoryPath:"",
                 addSubCategoryName:"",
                 imageName:"",
                 condition:{
-                  imgUpload: "methodFile"
+
+                    imgUpload: "methodFile",
+                    reviseComplete:'revise',
+                    reviseimgUpload:"methodFile"
                 },
                 uploadMethod: [
                     {text: '컴퓨터에서 찾기', value: 'methodFile'},
@@ -177,11 +235,14 @@
         },
         mounted: function () {
             this.getCategory(),
-            console.log(this)
+                console.log(this)
 
         },
         methods:{
-
+            revisebtnChange(){
+                this.condition.reviseComplete='complete'
+            },
+            //대분류 카테고리 클릭시
             categoryClicked(categoryName, categoryid){
                 this.selectedCategoryName= categoryName;
                 this.selectedCategoryID=categoryid;
@@ -194,68 +255,98 @@
                 this.formData.append('categoryFile', newFile[0], newFile[0].name);
             },
 
+            //카테고리 리스트 가져오기
             async getCategory(){
                 var categoryList  = await this.$http.get('http://localhost:3000/category');
+                console.log(categoryList.data);
                 this.mainCategoryList=categoryList.data;
             },
 
-             async addCategory(){
+            //카테고리 추가
+            async addCategory(){
                 var adminVue= this;
                 //https://picsum.photos/250/250/?image=54
                 var categoryData ={
                     name:this.addCategoryName,
                     img_path: this.addCategoryPath
                 };
-                 var categoryDataUpload = await this.$http.post('http://localhost:3000/category', categoryData)
-                 if(this.condition.imgUpload=='methodFile'){
-                     var imgFileUpload = await adminVue.$http.post('http://localhost:3000/category/files/'+categoryDataUpload.data._id, adminVue.formData,{ headers: { 'Content-Type': 'multipart/form-data' } })
-                     adminVue.mainCategoryList.push(imgFileUpload.data);
-                 } else {
-                     console.log("methodURL!")
+                var categoryDataUpload = await this.$http.post('http://localhost:3000/category', categoryData)
+                if(this.condition.imgUpload=='methodFile'){
+                    var imgFileUpload = await adminVue.$http.post('http://localhost:3000/category/files/'+categoryDataUpload.data._id, adminVue.formData,{ headers: { 'Content-Type': 'multipart/form-data' } })
+                    adminVue.mainCategoryList.push(imgFileUpload.data);
+                } else {
+                    console.log("methodURL!")
+                    adminVue.mainCategoryList.push(categoryDataUpload.data);
+                }
+                this.addCategoryName="";
+                this.addCategoryPath="";
 
-                     adminVue.mainCategoryList.push(categoryDataUpload.data);
-                 }
-                 this.addCategoryName="";
-                 this.addCategoryPath="";
 
-
-                 /*
-                this.$http.post('http://localhost:3000/category', categoryData)
-                .then((result) => {
-                    console.log(adminVue.formData.keys())
-                    return adminVue.$http.post('http://localhost:3000/category/files/'+result.data._id, adminVue.formData,{ headers: { 'Content-Type': 'multipart/form-data' } })
-                })
-                .then((result) => {
-                    //result.data.img_path=tmpImgPath
-                    adminVue.mainCategoryList.push(result.data);
-                })
+                /*
+               this.$http.post('http://localhost:3000/category', categoryData)
+               .then((result) => {
+                   console.log(adminVue.formData.keys())
+                   return adminVue.$http.post('http://localhost:3000/category/files/'+result.data._id, adminVue.formData,{ headers: { 'Content-Type': 'multipart/form-data' } })
+               })
+               .then((result) => {
+                   //result.data.img_path=tmpImgPath
+                   adminVue.mainCategoryList.push(result.data);
+               })
 */
             },
+
+
             addSubCategory(){
                 var adminVue= this;
                 var subcategoryData ={
                     name:this.addSubCategoryName
                 };
-
                 var addsubURL='http://localhost:3000/category/'+this.selectedCategoryID+'/sub';
                 var addedData= this.$http.post(addsubURL, subcategoryData);
                 addedData.then(function (result) {
                     for(let i=0;i<adminVue.mainCategoryList.length;i++){
                         if(adminVue.mainCategoryList[i]._id==adminVue.selectedCategoryID){
+                            console.log(result.data.sub_category);
                             adminVue.mainCategoryList[i].sub_category= result.data.sub_category
                         }
                     }
                     //adminVue.mainCategoryList.sub_category=result.data.sub_category;
-                   console.log(result.data);
+                    console.log(result.data);
 
                 });
                 this.addSubCategoryName=""
 
             },
 
-            categoryUpdate(data){
+            async categoryUpdate(data){
+                var adminVue= this;
+                this.selectedCategoryName=data.name;
+                var categoryData ={
+                    name:data.name,
+                    img_path: this.reviseCategoryPath
+                };
+
                 var updateURL='http://localhost:3000/category/edit/'+data._id;
-                var updatedCategory = this.$http.post(updateURL,data);
+                var updatedCategory = await this.$http.post(updateURL,categoryData);
+                if(this.condition.reviseimgUpload=='methodFile'){
+                    var imgFileUpload = await adminVue.$http.post('http://localhost:3000/category/files/'+data._id, adminVue.formData,{ headers: { 'Content-Type': 'multipart/form-data' } })
+                    console.log(imgFileUpload.data);
+                    for(let i=0;i<this.mainCategoryList.length;i++){
+                        if(this.mainCategoryList[i]._id==data._id){
+                            adminVue.mainCategoryList[i]=imgFileUpload.data
+                        }}
+
+                } else {
+                    console.log("Updated CategoryPath>>>"+this.reviseCategoryPath)
+                    for(let i=0;i<this.mainCategoryList.length;i++){
+                        if(this.mainCategoryList[i]._id==data._id){
+                            this.mainCategoryList[i].img_path = this.reviseCategoryPath;
+                        }}
+                    //adminVue.mainCategoryList.push(categoryDataUpload.data);
+                }
+
+                this.reviseCategoryPath="";
+                this.condition.reviseComplete='revise';
             },
 
             subcategoryUpdate(data){
@@ -266,14 +357,26 @@
 
 
             categoryDelete(id){
-                var deleteURL='http://localhost:3000/category/delete/'+id;
-                var deletedData= this.$http.delete(deleteURL);
-
                 for(let i=0;i<this.mainCategoryList.length;i++){
                     if(this.mainCategoryList[i]._id==id){
-                        this.mainCategoryList.splice(i,1);
+                        if(this.mainCategoryList[i].sub_category.length==0){
+                            var deleteURL='http://localhost:3000/category/delete/'+id;
+                            this.$http.delete(deleteURL);
+                            this.mainCategoryList.splice(i,1);
+                        } else{
+                            alert("세부 카테고리가 존재합니다.")
+                        }
+
                     }
                 }
+
+                /*
+                                for(let i=0;i<this.mainCategoryList.length;i++){
+                                    if(this.mainCategoryList[i]._id==id){
+
+                                        this.mainCategoryList.splice(i,1);
+                                    }
+                                }*/
             },
             subCategoryDelete(id){
                 var subdeleteURL='http://localhost:3000/category/'+this.selectedCategoryID+'/sub/delete/'+id;
@@ -295,8 +398,12 @@
 </script>
 
 <style scoped>
-
+    .hrLine20{
+        padding-left:10%;
+        padding-right:10%;
+    }
     .mainCategoryGroup{
+        color: #706f76;
         padding: 10px;
         background-color: #d2e8ff;
         float:left;
@@ -305,9 +412,12 @@
     .mainCategoryTitle{
         font-size: 20px;
         width: 200px;
+        margin-bottom: 5%;
     }
 
     .categoryTable{
+
+        width: 90%;
         margin-bottom: 15px;
         background-color: white;
 
@@ -324,10 +434,10 @@
     }
 
     .createCategory{
+        color: #706f76;
         background-color: white;
-        width: 410px;
+        width: 100%;
         border-radius:5px;
-        border: 3px solid #007bff;
         padding-top: 5px;
         margin-top: 10px;
         margin-bottom: 30px;
@@ -337,10 +447,9 @@
         margin-top: 15px;
         margin-bottom: 15px;
     }
-
     .plusbtn{
-        margin-top: 30px;
-        width: 407px;
+        margin-top: 10px;
+        width: 100%
     }
 
     .subplusbtn{
@@ -353,30 +462,44 @@
         float:left;
         alignment: left;
     }
+    .createsubCategoryArea{
+        padding-top: 20px;
+        padding-bottom: 20px;
 
+    }
+    .createsubCategory{
+        background-color: white;
+        width: 410px;
+        border-radius:5px;
+        border: 3px solid #007bff;
+        padding-top: 5px;
+        margin-top: 10px;
+        margin-bottom: 30px;
+        margin-left: 20%;
+    }
     .subCategoryGroup{
-        width: 1000px;
+        margin-left: 15%;
+        width: 700px;
         position: relative;
     }
 
     .subCategoryTitle{
         text-align: center;
-        margin-top: 100px;
-        margin-left: 30px;
+        color: #706f76;
+        margin-left: 20%;
         margin-bottom: 40px;
     }
     .subCategoryListTitle{
-        width:200px;
+        width:80%;
         margin-top:5px;
         margin-bottom:5px;
     }
     .subCategorySingle{
-        background-color: white;
-
+        background-color: #d2e8ff;
         width: 300px;
         border-radius:10px;
-        border: 5px solid #007bff;
-        padding-top: 5px;
+        padding:8px;
+
         margin-left: 20px;
         margin-bottom: 50px;
     }
